Phoenix webhook: cast the kobo parameters to bigint
An inbound application over about NGN 21.5m was rejected with a 500:

  unable to encode 5000000000 into binary format for int4 (OID 23):
  5000000000 is greater than maximum value for int4

amount_requested_kobo and monthly_income_kobo are bigint columns, so the column
was never the problem. COALESCE($9, 0) and NULLIF($11,0) let Postgres infer the
PARAMETER type from the bare literal 0, which is int4 — so the bind overflowed
before it ever reached the column. int4 tops out at 2,147,483,647 kobo, which is
NGN 21,474,836.47.

Casting $9 and $11 to bigint fixes the inference. $10 is tenor_months and is
genuinely small, so it stays int4.

This would not surface in ordinary use: it needs an application larger than
NGN 21.5m, so it was waiting for a big-ticket borrower. Found end-to-end testing
the inbound webhook after wiring the workspace to the live Phoenix deployment.

Verified after the fix: application.created lands at NGN 50m, a replayed event_id
returns duplicate:true, a tampered signature is still rejected 401, and
decision.completed populates eye_score/eye_rating/dti_pct/decision.

+++ backend-go/handlers/phoenix.go
@@ -708,7 +708,7 @@ func phoenixUpsertApplication(ctx context.Context, db *core.DB, pa phoenixApplic
 			-- table with no default, so an insert MUST supply them. A Phoenix payload
 			-- that omits product_type would otherwise blow up on the constraint.
 			COALESCE(NULLIF($8,''), 'Unspecified'),
-			COALESCE($9, 0), COALESCE($10, 0), NULLIF($11,0),
+			COALESCE($9::bigint, 0), COALESCE($10, 0), NULLIF($11::bigint, 0),
 			NULLIF($12,''), NULLIF($13,''), $14, $15,
 			COALESCE($16::timestamptz, NOW()), 'not_required', NOW(), NOW(), NOW()
 		)
@@ -730,9 +730,9 @@ func phoenixUpsertApplication(ctx context.Context, db *core.DB, pa phoenixApplic
 			applicant_email       = COALESCE(NULLIF($6,''),  app.loan_applications.applicant_email),
 			employer              = COALESCE(NULLIF($7,''),  app.loan_applications.employer),
 			product_type          = COALESCE(NULLIF($8,''),  app.loan_applications.product_type),
-			amount_requested_kobo = COALESCE(NULLIF($9,0),   app.loan_applications.amount_requested_kobo),
+			amount_requested_kobo = COALESCE(NULLIF($9::bigint,0),   app.loan_applications.amount_requested_kobo),
 			tenor_months          = COALESCE(NULLIF($10,0),  app.loan_applications.tenor_months),
-			monthly_income_kobo   = COALESCE(NULLIF($11,0),  app.loan_applications.monthly_income_kobo),
+			monthly_income_kobo   = COALESCE(NULLIF($11::bigint,0),  app.loan_applications.monthly_income_kobo),
 			sector_code           = COALESCE(NULLIF($12,''), app.loan_applications.sector_code),
 			purpose               = COALESCE(NULLIF($13,''), app.loan_applications.purpose),
 			status                = COALESCE(NULLIF($14,''), app.loan_applications.status),
